Change ufunc_list to an avl tree.
Take advantage of that and sorted built in funcs to speed up tab
completion.

Fix several memory leaks/errors.

+++ narray.h
@@ -19,7 +19,7 @@ void* Narray_delete(Narray* a, const char* key);
 void* Narray_remove(Narray* a, int index);
 int Narray_find(Narray* a, const char* key, void** data);
 int Narray_replace(Narray* a, int i, void* New, void** old);
-int Narray_insert(Narray* a, const char* key, void* data, int pos);
+int Narray_insert(Narray* a, const char* key, void* data, size_t pos);
 int Narray_get(const Narray* a, const int i, char** key, void** data);
 int Narray_count(const Narray* a);
 void Narray_free(Narray* a, Narray_FuncPtr fptr);
